Routed mousewheel events to sliders in options.scalar widgets Custodian now has IsCtrlPressed to facilitate Ctrl+Mouse operations

+++ source/rococo/rococo.gui.retained/rococo.gr.scrollable-menu.cpp
@@ -67,7 +67,7 @@ namespace GRANON
 
 			options.push_back({ name, caption, button });
 
-			viewport->SetDomainHeight(options.size() * 48);
+			viewport->SetDomainHeight((int) (options.size() * 48));
 		}
 
 		EGREventRouting OnCursorClick(GRCursorEvent& ce) override
